requirements.txt
# Backend Dependencies - Sistema IBV
# Django y REST Framework
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.28.0

# Servidor de producción
gunicorn>=21.2.0

# Base de datos
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# CORS para frontend
django-cors-headers>=4.3.0

# Generación de reportes
pandas>=2.2.0
openpyxl>=3.1.2

# Generación de QR
qrcode[pil]>=7.4.2

# Manejo de imágenes
Pillow>=10.2.0

# Supabase
supabase>=2.3.0

# Variables de entorno
python-decouple>=3.8

# Utilidades
python-dotenv>=1.0.0
pytz>=2024.1
